Fix build error reported on Arduino with BT816/7 encoding CMD_BITMAP_SOURCE2.
Typo in EVE_config.h comment on DP-0501-11A panel.
Touchscreen test forces a calibration on startup to enable buttons if previously stored calibration is incorrect.

+++ source/EVE_API.c
@@ -747,7 +747,7 @@ void EVE_BITMAP_SOURCE(int32_t addr)
 #if IS_EVE_API(3, 4)
 void EVE_BITMAP_SOURCE2(uint8_t flash_or_ram, int32_t addr)
 {
-    HAL_Write32(EVE_ENC_BITMAP_SOURCE2(flash_or_ram, (int32_t)addr));
+    HAL_Write32(EVE_ENC_BITMAP_SOURCE2((uint32_t)flash_or_ram, (int32_t)addr));
     HAL_IncCmdPointer(4);
 }
 #endif
